Added a unique constraint for Hotel id & name together
I noticed that in the API data there are multiple instances of hotels with the same name and the same city (but with a different id).
I added a 'unique together' constraint in the models.
NOTE: I would never do this if other people were already using this project, since the DB tables need to be removed and installed from scratch,
but here I know no one else is using the project.

- Added unique together constraint for Hotel id & name
- Deleted migration files and added a new migration file
- Added an IntegrityError to the fetch_api_data.py file to catch and log instances of duplicate hotels